[UPDATE] `init_database` to implement new functions
- populate_permissions
- populate_roles
- populate_users

+++ app/database/init_database.py
@@ -1,26 +1,22 @@
 from sqlmodel import SQLModel
 
+from app.core import settings
 from app.models import *
 from app.database import engine
-from app.database import roles as first_roles
-from app.database import superuser
-from app.resources import roles
-from app.resources import users
+from app.database.populate import populate_permissions
+from app.database.populate import populate_roles
+from app.database.populate import populate_users
 
-
-def create_db_and_tables():
-    SQLModel.metadata.create_all(engine)
-    print("DEBUG:   ", "Database and tables created.")
-
-def drop_db_and_tables():
-    SQLModel.metadata.drop_all(engine)
-    print("DEBUG:   ", "Database and tables dropped.")
-
-def create_first_roles():
-    for role in first_roles:
-        roles.create(role)
-    print("DEBUG:   ", "First roles created.")
-
-def create_first_admin():
-    users.create(superuser)
-    print("DEBUG:   ", "First administrator created.")
+def init_database()->None:
+    """
+    Run the database initialization functions according to the settings.
+    """
+    if settings.DB_RESET:
+        SQLModel.metadata.drop_all(engine)
+        print("DEBUG:   ", "Database and tables dropped.")
+        SQLModel.metadata.create_all(engine)
+        print("DEBUG:   ", "Database and tables created.")
+    if settings.DB_POPULATE:
+       populate_permissions(settings.DB_POPULATION) 
+       populate_roles(settings.DB_POPULATION) 
+       populate_users(settings.DB_POPULATION)
